+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Changelog
 **Changed**
 
 - Rename ``strategies.query`` to ``strategies.queries`` and ``strategies.schema`` to ``strategies.schemas``, so they conform with the recommended naming of Hypothesis strategies.
+  Old names are preserved for backward-compatibility.
 
 `0.5.1`_ - 2021-08-05
 ---------------------